refactor: 소켓 서비스 분리
